requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.6
motor==3.6.0
openai==1.57.4
pydantic==2.10.4
pymongo==4.9.2
python-dotenv==1.0.1
pytrends==4.9.2
requests==2.32.3
uvicorn[standard]==0.34.0
